pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "apscheduler>=3.11.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "python-telegram-bot>=21.11.1",
]
